Add support for plugns and convert axios hook to plugin
