+++ tsconfig.json
@@ -9,6 +9,7 @@
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "noEmit": true,
+    "skipLibCheck": true,
     "target": "ES5",
     "module": "CommonJS",
     "moduleResolution": "Node",
